Reorg imaging into analysis + tools and split out persistence
Splits the overloaded `imaging/` module into purpose-aligned subtrees
and lifts FITS I/O, the image cache, and exposure-document storage
into a new top-level `persistence/` module:

- `imaging/analysis/` — pure single-purpose kernels (background, stars,
  hfr, fwhm, snr, stats, pixel). Generic over `Pixel`, `ArrayView2`-only,
  no async, no I/O — unit-testable without a runtime.
- `imaging/tools/` — compositional analyzers (measure_basic,
  measure_stars). Future home for `auto_focus` and `center_on_target`.
- `persistence/` — `cache.rs`, `fits.rs`, and `document.rs` (moved from
  the crate root). Future filename-template / token-resolver work
  attaches here, not under `imaging/`.

`imaging/mod.rs` re-exports the flat `crate::imaging::<symbol>` shape
so MCP wiring is unaffected; persistence symbols (`ImageCache`,
`CachedPixels`, `CachedImage`, `ExposureDocument`, `write_fits`, etc.)
move to `crate::persistence::*` with caller updates.

Lands ahead of Phase 6 so `auto_focus` / `center_on_target` arrive on
the new layout. No behavior change. 162 lib tests + 139 BDD scenarios
green; clippy -D warnings clean. Docs updated: rp.md Module Structure
and plan Phase 6 prep section.
