+++ services/rp/src/imaging/analysis/background.rs
