+++ services/rp/src/imaging/analysis/pixel.rs
