+++ services/rp/src/imaging/analysis/snr.rs
