+++ services/rp/src/imaging/analysis/stars.rs
